Add: 특정 회원의 측정 기록들(목록) 가져오기
