--- models/records_user.js
@@ -1,7 +1,0 @@
-const myDataSource = require('./init');
-
-const getUserRecords = async userId => {
-    return await myDataSource.query(
-    );
-  };
-  module.exports = { getUserRecords };
