+++ routes/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const recordsUserGetRouter = require('./records_user');
+const recordsUserGetRouter = require('./records');
 
 router.use(recordsUserGetRouter);
