+++ routes/records.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const { getUserRecords } = require('../controllers/records');
+
+router.get('/records/:userId', getUserRecords)
+
+module.exports = router;
