--- routes/records_user.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const { getUserRecords } = require('../controllers/records_user');
-
-router.get('/records', getUserRecords)
-
-module.exports = router;
