+++ services/records.js
@@ -1,8 +1,8 @@
-const recordsModels = require('../models/records_user');
+const recordsModels = require('../models/records');
 const getUserRecords = async userId => {
   const res = await recordsModels.getUserRecords(userId);
 
-  res[0].detail_list = JSON.parse(res[0].detail_list);
+  res[0].measurement_data = JSON.parse(res[0].measurement_data);
   return res;
 };
 module.exports = { getUserRecords };
